move router to component repo
