+++ packages/serverless-nextjs-component/__tests__/fixtures/manifest.json
